Reorg code and update the filter a bit
